moved all gc non allocation stuff to own header and source files
